Math.Net library v0.17
Math.Net is a symbolic math library for .Net. For further information see 'www.cdrnet.net/projects/nmath/' or 'nmath.sourceforge.net/wiki/index.php/Main_Page' for the currently developed version. Documentation of the library can be found in the 'nmath.reference-0.17.chm' helpfile in the root directory of the project. Anyway, to get an impression what the library is capable of have a look at the Demos on the projects homepage.

The project is currently undergoing heavy changes, so I decided to make a snapshot of the version I'm using for my Palladio.Reliability library.
I also made some slight changes:
- added a property Variables to the VariableManager, so that it is possible to get  the variables used by an expression
- added a property Data to the ValueMatrix to get the internal double[,] array


git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code@192 769cc9a0-fb0c-0410-8c2f-a2305f0e6268