we noticed that just copying all files into one folder does not work if the original models are spread over several folders or plugins. I have now updated CreateWorkingCopyOfModelsJob to create a folder structure within the generated plugin, e.g.
models
  platform
    resource
      projectA
        default.repository
        default.system
      projectB
        default.allocation
        default.resourceenvironment

Looks a bit verbose (maybe we want  to remove the first two levels, but not sure whether we always have platform/resource as the scheme and prefix of all model URIs). 

It probably also only works if the model files have relative paths to each other, such as in the example default.allocation 

  ../projectA/default.system

I doubt that it would work if default.allocation refers to a path 

  platform:/resource/projectA/default.system

Does anyone know whether we always have relative paths in the models?

git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Core/trunk@25298 769cc9a0-fb0c-0410-8c2f-a2305f0e6268